requirements.txt
flask
mysql-connector-python
pymongo 
bcrypt
sqlalchemy
secure-smtplib
pyotp
python-dotenv
